Working on PersistenceManager Tests.

+++ de.woq.osgi.akka.persistence/src/main/scala/de/woq/osgi/akka/persistence/internal/PersistenceBackend.scala
@@ -23,7 +23,7 @@ import akka.event.LoggingAdapter
 
 trait PersistenceBackend {
   def initBackend(baseDir: String, config: Config)(implicit log: LoggingAdapter) : Unit
-  def store(obj : DataObject)(implicit log: LoggingAdapter) : Int
+  def store(obj : DataObject)(implicit log: LoggingAdapter) : Long
   //def get(uuid: String)(implicit log: LoggingAdapter) : DataObject
   def shutdownBackend()(implicit log: LoggingAdapter) : Unit
 }
